+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 27. Add `totalCount` parameter to `tableOptions` prop for `PagedTable`. This allows the display of the total number of results in table if it is known.
 28. Fix bug related to table pagination. For example when showing element 61 to 70 and user changes count per row to 25 per page, we now will show elements 51 to 75.
 29. Rename package to `@grailbio/components`, publish to github and npm.
+30. Fix bug to pass down onFocus and onBlur props in `TextInputV2` component.
 
 ## 1.0.0 - 2019-06-17